+++ modules/history/src/test/java/hosh/modules/history/HistoryModuleTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package hosh.modules;
+package hosh.modules.history;
 
 import hosh.spi.ExitStatus;
 import hosh.spi.InputChannel;
